Stage 1
Some executable script you want to make into a tool. This example is
clearly contrived, but it has all the bits we care about. We have some
arguments that are going to change every time we run it, some
configurable bits which probably don't change, and some arguments we
might want to default, but that could also be passed in.